+++ tests/EmptyTest.php
@@ -8,11 +8,11 @@
  * @copyright (c) 2017, Sebastian Rapetti
  * @license http://opensource.org/licenses/MIT MIT License
  */
-$injectionsRules = [
-    '\Linna\Storage\PdoStorage' => [
-        0 => $options['pdo_mysql']
-    ],
-    '\Linna\Auth\Password' => [
-        0 => $options['password'],
-    ],
-];
+use PHPUnit\Framework\TestCase;
+
+class EmptyTest extends TestCase
+{
+    public function testAnything(){
+        $this->assertEquals(true, true);
+    }
+}